Fix bug with distill sheets not in right dir for binding
The distill sheets were in the assets dir, and then used in bin dir
distill.py as `Path(__file__).parent.parent / "assets/forms/distill_sheets"`
which pointed to the relative path for assets dir. But when using singularity
or other containerization env, bin dir got binded, which means assets dir
wasn't inside the file system for the container. So either we could
pass the sheets directly what we did before, or move them into the bin dir.
See #431. I would like to keep them in the bin directory, to allow
specifying the sheets by name and not path. Removing the need to put the
paths in the config. Which more easily allows DRAM2 python code to someday
become a stand alone tool again that is wrapped in NF. Users who want to
use a custom carbon sheet or other sheet, can now just not specify carbon
and pass it in as a custom sheet instead of a custom carbon sheet path.