bug: Fix TRNA_COLLECT and COMBINE_ANNOTATIONS for large # inputs
rewrite TRNA_COLLECT to use pandas vectorized functions instead of
embedded for loops to significantly streamline creation of
collected_trnas.tsv with large # of inputs. Now instead of taking
hours or days, it will run in seconds or minutes.

rewrite COMBINTE_ANNOTATIONS to take directories of inputs instead
of a cli list of files so that when you have thousands and
thousands of mags or assemblies you don't run into your system's
ARG_MAX.