fix: Allow calling quast with arbitrary large no of inputs
Quast requires passing inputs as a space seperated list of files,
instead of a dir or other methods. This can run into ARG_MAX limits
So change it to passing a glob, such as `*.fa` instead. This should
fix the problem. We might need to move to batching at some point
for performance reasons.