feat: Rename ID col in GFF file
In the Prodigal GFF file, the metadata ID field is a generated
unique ID that is in the format 1_1, 1_2, 2_1, 2_2, etc.
This is a problem if people concatenate all the GFFs together since
then the unique IDs aren't unique. In DRAM1, the IDs were repalced
with the SeqID_Genenumber. So that is what we are doing here

We also replaced the python script that parsed the GFF into a summary
TSV for later use in DRAM2 into a tsv and replaced with with bash
parsing. Which benchmarking showed to be around 10-50 times faster.