fix: hmm parsing now takes top hit consistently
In the translation from DRAM1->2 and moving to nextflow,
the code to take the tophit for the hmm search had a couple
bugs in it depending on the hmm db. It would sometimes
take just the first hit outputted in the data sheet,
it would calcuate the bitscore wrong, and it wouldn't
do the info sheets right (since it was calculating bitscore
wrong).
Now it reports both full sequence bitscore and domain bitscore.
It takes the top it and also reports all other hits as well.
All of this is done in a generic way so that adding generic hmm
support will be trivial for next release after we fix this bug
Also add support for metals hmm DB.